Start is fractional coordinates

+++ stubs/Pillow/PIL/ImageFont.pyi
@@ -74,7 +74,7 @@ class FreeTypeFont:
         stroke_width: float = ...,
         anchor: str | None = ...,
         ink=...,
-        start: tuple[int, int] | None = ...,
+        start: tuple[float, float] | None = ...,
     ): ...
     def getmask2(
         self,
@@ -87,7 +87,7 @@ class FreeTypeFont:
         stroke_width: float = ...,
         anchor: str | None = ...,
         ink=...,
-        start: tuple[int, int] | None = ...,
+        start: tuple[float, float] | None = ...,
         *args,
         **kwargs,
     ): ...
